Remove duplicates from non/cron jobs
